ok nå er det bare auth som screwer up med hook error
