+++ marry_monio_frontend/src/features/Auth/Layout.sass
@@ -1,0 +1,10 @@
+@use "src/index" as *
+
+    
+main
+    margin-top: 5vh
+    margin-right: 15%
+    margin-left: 15%
+    height: 100%
+    padding: 10vh
+    background-color: $secoundary-bg-color
